# static/swagger.yaml
swagger: "2.0"
info:
  title: Password Keeper
  version: 1.0.0
  description: >-
    API for managing users, roles, folders, passwords, password history, audit logs,
    and security questions. 
    

    All protected endpoints require a bearer token in the Authorization header. Use the "/user/login" endpoint to obtain a token, then include it as "Bearer <token>" for authenticated requests.


    How to use this API:
      1. Create a user with /user/.
      2. Sign in with /user/login to receive an auth token.
      3. Copy the token and add it to the Authorization header as "Bearer <token>".
      4. Call protected endpoints such as /folder/, /password/, /roles/, or /security/.
      5. If your token expires, use /user/refresh with a refresh token to obtain a new one.
host: 127.0.0.1:5000
schemes:
  - http
consumes:
  - application/json
produces:
  - application/json
securityDefinitions:
  bearerAuth:
    type: apiKey
    name: Authorization
    in: header
responses:
  BadRequestError:
    description: Malformed JSON or an invalid request body. The API may return a validation object for schema problems.
    schema:
      $ref: '#/definitions/Error'
  UnauthorizedError:
    description: Missing, expired, or invalid bearer token.
    schema:
      $ref: '#/definitions/Error'
...
  NotFoundError:
    description: Requested resource was not found.
    schema:
      $ref: '#/definitions/Error'
  ConflictError:
    description: The request conflicts with an existing resource or current state.
    schema:
      $ref: '#/definitions/Error'
  UnprocessableEntityError:
    description: Business rule validation failed. The response uses the standard error envelope with error_code and details.
    schema:
      $ref: '#/definitions/Error'
  ServerError:
    description: The server encountered an unexpected error while processing the request.
    schema:
      $ref: '#/definitions/Error'
paths:
  /user/:
    post:
      tags: [Users]
      summary: Create a user
      description: Creates a new user account. Use this endpoint to register a new user before authenticating.
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/UserCredentials'
      responses:
        '201':
          description: User created
          schema:
            $ref: '#/definitions/User'
          examples:
            application/json:
              user_id: 1
              username: demoUser
              first_name: Demo
              last_name: User
              email: demo@example.com
              role: user
        '400':
          $ref: '#/responses/BadRequestError'
        '409':
          $ref: '#/responses/ConflictError'
        '422':
          $ref: '#/responses/UnprocessableEntityError'
        '500':
          $ref: '#/responses/ServerError'
    get:
      tags: [Users]
      summary: Get the current user
      description: Returns the authenticated user profile. This endpoint requires a valid bearer token.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User profile returned
          schema:
            $ref: '#/definitions/User'
        '401':
          $ref: '#/responses/UnauthorizedError'
    put:
      tags: [Users]
      summary: Update the current user
      description: Updates the authenticated user's profile. Send the full updated user payload in the request body.
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/User'
      responses:
        '201':
          description: User updated
          schema:
            $ref: '#/definitions/User'
        '400':
          $ref: '#/responses/BadRequestError'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '404':
          $ref: '#/responses/NotFoundError'
        '500':
          $ref: '#/responses/ServerError'
    delete:
      tags: [Users]
      summary: Delete the current user
      description: Deletes the authenticated user account.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User deleted
          schema:
            $ref: '#/definitions/DeleteResponse'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '404':
          $ref: '#/responses/NotFoundError'
        '500':
          $ref: '#/responses/ServerError'
...
  /password/:
    post:
      tags: [Passwords]
      summary: Create a password
      description: Stores a new password entry for the authenticated user. The payload includes the credential details and optional folder assignment.
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/PasswordCredentials'
      responses:
        '201':
          description: Password created
          schema:
            $ref: '#/definitions/Password'
          examples:
            application/json:
              password_id: 42
              folder_id: 10
              user_id: 1
              password_name: GitHub
              username: demoUser
              email: demo@example.com
              encripted_password: 8f9a6d2c4e7b1a3d
        '401':
          $ref: '#/responses/UnauthorizedError'
        '422':
          $ref: '#/responses/UnprocessableEntityError'
        '500':
          $ref: '#/responses/ServerError'
    get:
      tags: [Passwords]
      summary: List passwords
      description: >-
        Returns password entries for the authenticated user, including any folder association.
        Use the optional query parameters limit and offset to page through results.
      security:
        - bearerAuth: []
      parameters:
        - name: limit
          in: query
          required: false
          type: integer
          default: 50
          description: Maximum number of password records to return. The server caps this at 100.
        - name: offset
          in: query
          required: false
          type: integer
          default: 0
          description: Number of password records to skip before returning the next page.
      responses:
        '200':
          description: Passwords returned
          schema:
            type: array
            items:
              $ref: '#/definitions/Password'
          examples:
            application/json:
              - password_id: 42
                folder_id: 10
                user_id: 1
                password_name: GitHub
                username: demoUser
                email: demo@example.com
                encripted_password: 8f9a6d2c4e7b1a3d
              - password_id: 43
                folder_id: 11
                user_id: 1
                password_name: Azure
                username: demoUser
                email: demo@example.com
                encripted_password: 1a2b3c4d5e6f7g8h
        '401':
          $ref: '#/responses/UnauthorizedError'
    put:
      tags: [Passwords]
      summary: Update a password
      description: Updates an existing password record for the authenticated user. This also updates the password history trail for the entry.
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Password'
      responses:
        '201':
          description: Password updated
          schema:
            $ref: '#/definitions/Password'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '404':
          $ref: '#/responses/NotFoundError'
        '422':
          $ref: '#/responses/UnprocessableEntityError'
        '500':
          $ref: '#/responses/ServerError'
    delete:
      tags: [Passwords]
      summary: Delete a password
      description: Deletes a password record for the authenticated user.
      security:
        - bearerAuth: []
      parameters:
        - in: body
          name: body
          required: true
          schema:
            $ref: '#/definitions/Password'
      responses:
        '200':
          description: Password deleted
          schema:
            $ref: '#/definitions/DeleteResponse'
        '401':
          $ref: '#/responses/UnauthorizedError'
        '404':
          $ref: '#/responses/NotFoundError'
        '422':
          $ref: '#/responses/UnprocessableEntityError'
        '500':
          $ref: '#/responses/ServerError'
...
definitions:
  User:
    type: object
    example:
      user_id: 1
      username: demoUser
      first_name: Demo
      last_name: User
      email: demo@example.com
      role: user
      create_date: '2026-08-10T12:00:00Z'
      updated_date: '2026-08-10T12:30:00Z'
    properties:
      user_id:
        type: integer
      username:
        type: string
      password:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      email:
        type: string
      role:
        type: string
      create_date:
        type: string
        format: date-time
      updated_date:
        type: string
        format: date-time
  UserCredentials:
    type: object
    required: [username, password, first_name, last_name, email]
    example:
      username: demoUser
      password: strongPassword123
      first_name: Demo
      last_name: User
      email: demo@example.com
    properties:
      username:
        type: string
      password:
        type: string
      first_name:
        type: string
      last_name:
        type: string
      email:
        type: string
        format: email
...
  Password:
    type: object
    required: [user_id, email, encripted_password]
    example:
      password_id: 42
      folder_id: 10
      user_id: 1
      password_name: GitHub
      username: demoUser
      email: demo@example.com
      encripted_password: 8f9a6d2c4e7b1a3d
      created_date: '2026-08-10T12:10:00Z'
      last_updated_date: '2026-08-10T12:10:00Z'
    properties:
      password_id:
        type: integer
      folder_id:
        type: integer
        x-nullable: true
      user_id:
        type: integer
      password_name:
        type: string
      username:
        type: string
      email:
        type: string
        format: email
      encripted_password:
        type: string
      created_date:
        type: string
        format: date-time
      last_updated_date:
        type: string
        format: date-time
  PasswordCredentials:
    type: object
    required: [user_id, email, encripted_password]
    example:
      folder_id: 10
      user_id: 1
      password_name: GitHub
      username: demoUser
      email: demo@example.com
      encripted_password: 8f9a6d2c4e7b1a3d
    properties:
      folder_id:
        type: integer
        x-nullable: true
      user_id:
        type: integer
      password_name:
        type: string
      username:
        type: string
      email:
        type: string
        format: email
      encripted_password:
        type: string
...
  DeleteResponse:
    type: object
    example:
      message: Resource removed successfully
    properties:
      message:
        type: string
  Error:
    type: object
    example:
      status: error
      message: Password not found
      error_code: password_not_found
      details:
        code: password_not_found
        domain: password
        operation: lookup
        message: The requested password could not be found.
    properties:
      status:
        type: string
        description: Always "error" for failed responses.
      message:
        type: string
        description: Human-readable explanation of the failure.
      error_code:
        type: string
        description: Stable machine-readable code for the error category.
      details:
        type: object
        description: Structured information describing the domain, operation, and detail message for the failure.
        properties:
          code:
            type: string
          domain:
            type: string
          operation:
            type: string
          message:
            type: string